Politician_Quarterly_Analysis/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
quiverquant>=0.2.1
python-dotenv>=0.19.0